% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mfi_population}
\alias{mfi_population}
\title{Population by age, sex, and country}
\format{
A data frame with 458,304 rows and 7 columns:
\describe{
\item{Country Name}{Country name.}
\item{Indicator Name}{World Bank indicator name.}
\item{Year}{Year of observation (factor).}
\item{value}{Population count.}
\item{age}{Age group.}
\item{unit}{Unit of measurement.}
\item{sex}{"Male" or "Female".}
}
}
\source{
World Bank Open Data (\url{https://data.worldbank.org}).
}
\usage{
mfi_population
}
\description{
Long-format population counts by five-year age group and sex, used to
drive the population pyramid module.
}
\examples{
\donttest{
head(mfi_population)
}
}
\keyword{datasets}
