% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_index_table.R
\name{mod_index_table_ui}
\alias{mod_index_table_ui}
\title{Indicator index table module - UI}
\usage{
mod_index_table_ui(id)
}
\arguments{
\item{id}{Character. Module namespace id.}
}
\value{
A \code{DT::dataTableOutput} Shiny UI element.
}
\description{
Renders a DataTable output listing all available indicators.
}
\examples{
\dontrun{
shiny::fluidPage(mod_index_table_ui("index_table"))
}
}
\keyword{internal}
